+++ src/modules/users/infra/http/routes/users.routes.ts
@@ -20,10 +20,11 @@ usersRoutes.post('/updateEmail', usersController.updateEmail);
 usersRoutes.delete('/deleteUser', usersController.deleteUser);
 usersRoutes.get('/listUsers', usersController.listUsers);
 // usersRoutes.get('/getcalendar', usersController.getcalendar);
-usersRoutes.get('', usersController.getTokens);
+usersRoutes.get('/', usersController.getGoogleTokens);
+usersRoutes.get('/outlook', usersController.getOutlookTokens);
 usersRoutes.post('/getGoogleAuthUrl/', usersController.getGoogleAuthUrl);
 usersRoutes.post('/getOutlookAuthUrl/', usersController.getOutlookAuthUrl);
-usersRoutes.get('/getTokens', usersController.getTokens);
+// usersRoutes.get('/getTokens', usersController.getTokens);
 usersRoutes.post('/createEvent', usersController.createEvent);
 usersRoutes.post('/getRecommededTimes', usersController.getRecommendedTime);
 usersRoutes.get('/getEvents', usersController.getEvents);
